Refactored websockect consumer module to apps.consumer
